Move class CustomAmountUIModel.kt under order package
