+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/orders/creation/MapFeeLineToCustomAmountUiModel.kt
@@ -1,6 +1,7 @@
 package com.woocommerce.android.ui.orders.creation
 
 import com.woocommerce.android.model.Order
+import com.woocommerce.android.ui.orders.CustomAmountUIModel
 import com.woocommerce.android.ui.payments.customamounts.CustomAmountsDialog.Companion.CUSTOM_AMOUNT
 import javax.inject.Inject
