+++ WooCommerce/src/main/res/navigation/nav_graph_order_creations.xml
@@ -169,7 +169,7 @@
         android:label="CustomAmountsDialog">
         <argument
             android:name="customAmountUIModel"
-            app:argType="com.woocommerce.android.ui.orders.creation.CustomAmountUIModel"
+            app:argType="com.woocommerce.android.ui.orders.CustomAmountUIModel"
             android:defaultValue="@null"
             app:nullable="true"/>
     </dialog>
